requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
neo4j==5.14.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
PyJWT==2.8.1
pydantic-settings==2.1.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
